separate testing of client-sqs and client-s3 to separate test files so that TAV config can target one at a time
