+++ plugins/node/opentelemetry-instrumentation-aws-sdk/.tav.yml
@@ -1,5 +1,5 @@
 # Note: tests must set `SKIP_TEST_IF_DISABLE=true` to override usage of
-# `mocha --require '../../../scripts/skip-test-if.js' ...`.
+# `mocha --require '../../../scripts/skip-test-if.js' ...` if calling `npm test`.
 
 "aws-sdk":
   env:
@@ -9,7 +9,7 @@
     include: "^2.308.0"
     mode: max-7
   commands:
-    - npm run test
+    - mocha --require '@opentelemetry/contrib-test-utils' test/aws-sdk-v2.test.ts
 
 # Versions [3.363.0, 3.377.0] of all @aws-sdk/client-* were bad releases. See:
 # - https://github.com/open-telemetry/opentelemetry-js-contrib/pull/2464#issuecomment-2403652552
@@ -29,19 +29,25 @@
         include: "^3.6.1"
         exclude: "3.529.0 || >=3.363.0 <=3.377.0"
         mode: "max-7"
-      commands: npm run test
+      commands:
+        - mocha --require '@opentelemetry/contrib-test-utils' test/aws-sdk-v3-s3.test.ts
+        - mocha --require '@opentelemetry/contrib-test-utils' test/s3.test.ts
     - node: "16"
       versions:
         include: ">=3.6.1 <3.723.0"
         exclude: "3.529.0 || >=3.363.0 <=3.377.0"
         mode: "max-7"
-      commands: npm run test
+      commands:
+        - mocha --require '@opentelemetry/contrib-test-utils' test/aws-sdk-v3-s3.test.ts
+        - mocha --require '@opentelemetry/contrib-test-utils' test/s3.test.ts
     - node: "14"
       versions:
         include: ">=3.6.1 <3.567.0"
         exclude: "3.529.0 || >=3.363.0 <=3.377.0"
         mode: "max-7"
-      commands: npm run test
+      commands:
+        - mocha --require '@opentelemetry/contrib-test-utils' test/aws-sdk-v3-s3.test.ts
+        - mocha --require '@opentelemetry/contrib-test-utils' test/s3.test.ts
 
 "@aws-sdk/client-sqs":
   env:
@@ -52,16 +58,23 @@
         include: "^3.24.0"
         exclude: ">=3.363.0 <=3.377.0"
         mode: "max-7"
-      commands: npm run test
+      commands:
+        - mocha --require '@opentelemetry/contrib-test-utils' test/aws-sdk-v3-sqs.test.ts
+        - mocha --require '@opentelemetry/contrib-test-utils' test/sqs.test.ts
     - node: "16"
       versions:
         include: ">=3.24.0 <3.723.0"
         exclude: ">=3.363.0 <=3.377.0"
         mode: "max-7"
-      commands: npm run test
+      commands:
+        - mocha --require '@opentelemetry/contrib-test-utils' test/aws-sdk-v3-sqs.test.ts
+        - mocha --require '@opentelemetry/contrib-test-utils' test/sqs.test.ts
     - node: "14"
       versions:
         include: ">=3.24.0 <3.567.0"
         exclude: ">=3.363.0 <=3.377.0"
         mode: "max-7"
-      commands: npm run test
+      commands:
+        - mocha --require '@opentelemetry/contrib-test-utils' test/aws-sdk-v3-sqs.test.ts
+        - mocha --require '@opentelemetry/contrib-test-utils' test/sqs.test.ts
+
